MXS-5191: Ignore failure to register variable handlers
The failure is expected if more than one cache filter is used in the
same service. The information is still useful so it can be logged on the
info level.

Added a debug assertion into Session::add_variable() to catch cases
where the format of the variable is incorrect.

Added a note about this limitation into the cache filter's
documentation.