MXS-5075 Allow the same module command to work with positional and key-value arguments
If the same command is implemented with both argument types, the
implementation is chosen depending on if the passed arguments
contain a key-value pair. To prevent confusion, such overlapping
commands must be designed such that a valid positional argument
should never have a value that looks like a key-value pair.