MXS-5197 Fix trigger drop SQL syntax for schema tree explorer
Right-click on a trigger node in the schema tree explorer, the auto-generated
SQL for dropping the trigger is invalid.

The right syntax is `DROP TRIGGER [schema_name.]trigger_name` instead of
`DROP TRIGGER [table_name.]trigger_name`

This adjusts the `qualified_name` for `TRIGGER` node to include
the schema name instead of the table name.