MXS-5136 Handle replacement projections in post-processing phase
When fetching documents, it is possible to specify what fields to
include, exclude or replace. Including and excluding a field can
be handled on the SQL-level, but replacement can in general not,
as the value to be used, may be calculated using an arbitrarily
complex expression.

With this change, if a projection only includes or excludes fields,
that will be handled on the SQL level. If there are replacements,
then the document will be fetched as such, and all actions will be
handled in a post-processing phase. This makes the $project stage
of the aggregate command actually useful.

The 'fields' specification of Find, FindAndModify, and the $project
stage in Aggregate are now handled in an identical fashion.