MXS-5188: Log last sd_notify time on fatal signals
Nearly always a SIGABRT for the MaxScale process is due to a systemd
watchdog timeout. By logging the notification interval and how many
seconds ago the last notification was done, we should be able to clearly
see whether the SIGABRT was due to the watchdog not being notified.

The information is useful and logging it for all signals could also
help. Mostly it'll be useful for SIGABRT.