MXS-5188: Make mxs2057_systemd_watchdog more reliable
By repeating the same filter several times with only one pattern and a
very long query, the time it takes for the query to be processed will be
much longer. This will cause PCRE2 errors about match limits and even
with the lower limits that were addd in newer versions of MaxScale, the
test should still fail as each namedserverfilter will process as much of
the query as possible and the repetition of the filters, not the
patterns, is what makes it slow.

Ideally, we'd use something that guarantees a stall in MaxScale but
without a programmable module like the luafilter, this is, luckily,
quite hard to do.