MXS-5185 Add `resultsetExtractor` module
This module is added to extract specific value from a resultset.
This adds `getDdl` function which will be later reused for showing
view, store procedure, function or trigger creation script in the
"ALTER" mode of the `DdlEditor`.

Additionally, this fixes a minor padding style issue caused by the
recent introduction of the `QueryResultTabWrapper` component.