MXS-5160: Fix postinstall script
The >& redirection into a file isn't supported by POSIX shells and it
only works with bash. As Debian and Ubuntu do not use bash for /bin/sh
during installation, this caused an invalid file descriptor error to be
emitted which caused the installation to fail.

Also ran shellcheck on the scripts and fixed a style warning about using
`$?` instead of using the command directly in the `if`.