Use system boost
The boost libraries on all supported systems support a new enough of a
version of boost for the Spirit x3 parsing to work.

There has been a change in Spirit x3 in Boost 1.77 that appears to have
caused the parsing to change slightly. The MASTER_GTID_WAIT() code and
the function definitions now require the sequence operators instead of
the expectation operators in order for alteration to work correctly. As
for whether the old behavior was a bug or not remains to be seen as no
Spirit x3 bug was found for this.