Use system libmemcached if available
The library appears to be available on nearly all supported operating
systems.  The exception is RHEL 9 where it's located in the Code Ready
Builder repository that isn't enabled by default which prevents us from
using it.

To handle the RHEL 9 case, libmemcached is built if CMake is configured
with BUNDLE=Y and the OS libraries do not have them.