Use system hiredis if available
The library is available on Debian and Ubuntu but on RHEL it's in EPEL
which means that it still has to be build on those systems.

The new BUNDLE option controls whether to download and build missing
dependencies if a version from the operating system was not found. This
is the default behavior but with -DBUNDLE=N the build can be configured
to instead skip modules that cannot be downloaded.