Bundle jwt-cpp version 0.6.0
This version of jwt-cpp is what's been used for a long time in MaxScale
and the included picojson in it makes it easy to include as a bundled
project. This removes the need to download it and as it's a header-only
library with a MIT license, including it in MaxScale should cause
minimal problems similar to how the INI library has been dealt with.