Include mariadb-connector-c as a subproject
This is how the MariaDB server does it and this is how MaxScale should
do it as well. This also makes it much easier to include the source code
in a source tarball as CPack will do this for submodules but won't do it
for projects added via ExternalProject_Add.