Remove obsolete files and make targets
The release notes and upgrading documents were hard-coded for the 1.2
release of MaxScale. This prevent builds from source tarballs from
working as they required perl to be present which is not listed as a
build dependency.

Removed the conversion of the changelog into a text file as the
conversion didn't really do anything and instad installed the raw
Markdown file.