MXS-5186 Handle the "Create table" option
Clicking on the mentioned option will open a new query tab where
the user can use the `TblEditor` component to create a new table,
similar to the "Alter Table" option.

This commit includes only the model creation part. The `TblEditor`
component will be adjusted later in another commit to support the
creation mode.