MXS-3952 Split SQL into statements, supporting delimiter changes
Reinstall the `dbgate-query-splitter` package as it supports custom
delimiters.

SQL will now be split using the `splitQuery` function from the mentioned
package. The result of that function includes trimmed position info,
which can be used to identify the actual delimiter of the statement.
So If the delimiter is not the default semicolon, the "LIMIT" clause
will not be injected.