MXS-5174 Extend $convert operator
Now jstests/aggregation/expressions/convert.js passes, but for
one case. The test case for converting a NumberLong(1) (64-bit)
to a date will fail, because the NumberLong(1) will turn into a
NumberInt(1) (32-bit) when the containing document is converted
from bson to json and then back to bson, and it is not permissible
to convert a NumberInt to a date. It may be possible to work around
that, but not just now.