Fix GCC 14 errors for external projects
The zlib library built by the connector was unnecessary as MaxScale
links against the system libraries. Now that it fails to compile, it is
better to simply disable it.

The tests and examples for the avro-c library also failed to compile and
they should've been disabled a long time ago as they're a waste of build
time.