MXS-4958: Use Yargs 16 in MaxCtrl
This allow Node.js version 10 to be used which is the oldest version
that's in the repositories for all supported systems. Rocky Linux 8 and
Ubuntu 20.04 both have it which means that downgrading Yargs makes it
possible to use the system Node.js on all operating systems.

Minor modifications were required for Node.js 10 to be usable with
MaxCtrl. Some of the functions used by MaxCtrl were introduced in
Node.js 14 but can be easily substituted with older alternatives or with
Lodash.