Allow alternative docker implementations in unit tests
Podman is more widely available compared to docker which is not
available in some of the repositories of the supported operating
systems. It also doesn't require root access unlike docker which
requires the user to be in the docker group. With DOCKER=podman and
DOCKER_COMPOSE=podman-compose, the unit tests can be run with podman.

On Ubuntu 20.04, podman is not available and the version of
docker-compose is old enough that it doesn't work at all which requires
the use of docker from Docker inc. which doesn't have docker-compose but
include compose as a command in docker. By using
DOCKER_COMPOSE='docker compose', the unit tests work with this version
of docker.