MXS-5207 Do not warn of the non-existence of /etc/maxscale.cnf.d
Do not warn of not reading a non-existing "$config-file.d". If it is
not there, it clearly will not be read.

Its presence is now logged with the following rationale: if a config
file that is expected to be read is not read, it will be pretty obvious,
but if a config file that is not expected to be read is read, it may
not be obvious but only appear as strange or unexpected behaviour.