MXS-4958: Use system Node.js by default
Now that MaxCtrl is compatible with Node.js 10, the one from the
operating system can be used directly. This removes the dependency on
pkg for the build.

Building MaxScale may still require a newer version of Node.js as the
GUI will use it to produce its output which is why the build scripts
still install Node.js 14 by default.

Also adjusted the minimum required version to be Node.js 10.