Remove installation of packages in build scripts
The build scripts installed the built package but the environment in
which it was installed was not clean. For example, nodejs was installed
from a tarball and the new dependency on it would not be met. Installing
the system nodejs on top of the already installed custom version would
meet the dependency but this results in a very broken system. There are
also development packages present from EPEL that may satisfy a
dependency when in a plain installation they would not be there.

There's also separate installation tests that cover the installation of
the generated packages which is why removing it from these builds
scripts should not degrade the test coverage.