Run git submodule update --init before testing
The system tests also need it as they use the MaxScale sources. The
copying of the sources also did not copy the hidden directories as the
scp command used shell globbing to select the files which doesn't match
the .git directory. Using rsync makes this easier as it copies the
contents without the need to know the source directory name.