Link to zlib in maxsql
The maxtest library failed to link as the order in which the
dependencies were resolved seemed to not be correct. This happens as the
connector-c now needs the symbols from zlib in order to work.

For a static target, the dependencies do not seem to be in the correct
order if the top level target (maxtest) links to the lower level target
(maxsql) and some a library that the lower level target needs is linked
to by the top level target.

The correct way to deal with this is to link to the necessary library in
the target that needs it so that the libraries are linked in the correct
order.