Make internal libraries OBJECT type libraries
This makes it so that they're built as static libraries even if external
compiler flags would create them as dynamic libraries. This is what
seems to happen when MaxScale is build from a source RPM on Fedora
40.

Making the libraries OBJECT libraries also reduces the amount of work
that has to be done at build time as the object files are linked
directly to whatever needs them instead of being bundled into a static
library.