Add workaround for Ubuntu 24.04 Bug #2023424
The following bug causes a misaligned access error from UBSAN if the
test_utils test is run with ASAN's stack-use-after-return detection.

https://bugs.launchpad.net/ubuntu/+source/gcc-13/+bug/2023424

The workaround to this is to disable the ASAN part for this test. The
test does not fail on Fedora 39 which has GCC 13.3.1 or on Fedora 40
which has GCC 14.2.1 and this appears to be a bug that's limited to
Ubuntu 24.04.