MXS-5225: Add maxctrl, maxkeys and maxpasswd manuals
Added manuals for the installed programs and generated the initial
versions with the help of the help2man program.

The MaxCtrl manual was amended with some sections from the Markdown
version and the formatting of it was cleaned up.

The manuals for maxkeys and maxpasswd are largely unmodified except the
removal of some auto-generated comments and sections.

The manuals were installed as executables which was not correct. The
manuals are read-only and thus do not need to be executable files.