MXS-5212 Fix number validation for `ParameterInput`
`int` and `duration` types accept negative numbers, and the API
returns these values as strings. Therefore, they must be parsed
and typecast to numbers. The validation rule has been modified
to check if the value is a number before further validation which
cause the validation to fail if the value is not typecast to number.