MXS-5138: Use "system maxctrl" in maxctrl tests
The unit tests for maxctrl now used the bundled source file and an
external Node.js process to run the commands if MAXCTRL_PROD=1 is is set
in the environment when running the `make test_maxctrl` target. This
makes it possible to verify that the final bundled source code passes
all tests while also allowing the testing to be done internally by the
test system for improved testing speed.

This also makes it possible to run the tests against other "maxctrl
compatible" processes. If MaxCtrl is to eventually to be rewritten in
another language, this makes it possible to run the test suite against
that program. This can be done by defining the program to run in the
MAXCTRL_CMD environment variable.