Fix source tarball creation on 24.02
The default package name was set to an undefined variable which caused
the package name to be empty when `make dist` was run. This was leftover
code from 23.08 that broke only the source tarball builds as they were
the only ones that used it. The PACKAGE_NAME variable was removed as a
part of the 24.02 reorganization of the packaging code.