MXS-5232: Allow DCB to be closed in more states
The states are now nearly the same as they are in track_query() which is
the function that accepts queries for processing. In all of the states
the protocol has advanced to a state where disconnecting it will not
increment the error counters in the server that lead to the MaxScale
host being blocked.