MXS-5206: Discard lagging connections
If a connection is behind by more than 300 second and the lag is larger
than twice the amount of max_replication_lag, the connection will be
closed when the next response from a backend arrives.

The use of m_check_stale means that only reads trigger the pruning of
the stale connections. This is a low overhead approach to the problem
and should be an acceptable compromise as max_replication_lag only makes
sense in scenarios where there are reads that get routed to replicas. It
also keeps the connection pruning code in one place.

The documentation includes the "algorithm" with which the limit is
calculated. The wording of the limit might not be obvious for all but a
pseudo-code explanation of the limit should help.