MXS-5215 Use nosql::Match in aggregation::Match
Now Match can be anywhere in the aggregation pipeline and
matching can either be performed so that the match is turned
into an SQL WHERE-clause or that the matching is done on
documents fetched from the database.

Next step is to add Query::Evaluator classes, so that the number
of operators that can be evaluated after documents have been
fetched are the same as the ones that can be turned into an SQL
condition.