MXS-5224: Add --archive option to `create report`
The new --archive option changes the output of the `create report` from
a JSON file to a tar archive. This allows the report to include files
from the local filesystem if they are readable by the current user. The
archive mode includes the same files that the maxscale_support_report.py
script includes and it also sanitizes the files by replacing all
passwords with placeholders.

Added a test case that makes sure the report generation works.