MXS-5241 Exclude the `SQL` option from the export options dropdown
In earlier versions, this option was disabled for the History and
Snippets tables. with changes in Vuetify's VSelect component, the
`disabled` attribute should now be part of the `props` attribute.
Nevertheless, instead of disabling the option, it would be better to
exclude it entirely, as data in the History and Snippets tables
cannot be exported as SQL since they are not results from SQL tables.