MXS-4785: Avoid useless sprintf usage
The output is only used if log_info is enabled. The calls can be
directly replaced with MXB_INFO calls as the only thing that was done
with it was to log it on the info level. Changed the overflow checks to
be assertions as now the info messages are logged immediately.