MXS-4785: Optimize regex matching in replicator
The query matching constructed a mxb::Regex object and then discarded it
immediately. Using the same instance again removes the need to compile
the pattern more than once.

The code was also pessimistic as it assumed a comment was present and
always re-assigned the return from replace() to the string.

This code would be run more often for workloads that used non-XA capable
engines where the commit is done as a "COMMIT" Query_event instead of an
Xid_event.