Use UPDATE_COMMAND for libmarias3
CMake 3.16 on Ubuntu 20.04 seems to not work correctly when
PATCH_COMMAND is used with FetchContent_Declare(): if any PATCH_COMMAND
is defined, the cloning of the git repository will fail as the repo name
will be empty or otherwise mangled.

Using the UPDATE_COMMAND should be an acceptable alternative as the
operation that's done is idempotent.