MXS-5246: Hex-encode non-UTF8 data in kafkacdc
Similar to MXS-5038, the use of the replicated binary and string data is
unsafe. If the data cannot be represented as UTF8, the string values
would be ignored and the binary values would result in malformed
JSON. The same solution that is taken in the avrorouter is used here as
well: if the data cannot be encoded as UTF8, it is encoded as a
hexadecimal string.