Disable LTO for binlogrouter with Boost 1.66
Using LTO with the binlogrouter and Boost 1.66 causes the link-time
optimization to end up with an internal compiler error. So far there
does not seem to be a way to avoid it by modifying the code and the only
option is to disable it.

This should only affect Rocky Linux 8 as it has the oldest version of
Boost.