MXS-5238: Add redirect_url to listeners
Added a new `redirect_url` parameter to listeners that complements the
existing `connection_metadata` feature. With this, clients can be more
easily redirected by defining the `redirect_url` parameter at runtime:

  maxctrl alter listener <name> redirect_url=<URL>

Added documentation for the new parameter and also explained the
behavior of the backend protocol when it comes to redirect_url values
read from the databases.

Extended the mxs4635_custom_sysvars test to cover the use of
redirect_url.