MXS-5256: Fix parsing of multi-value SET statements
The SET statement parsing in TrxBoundaryParser only expected one value
to be passed to it. In reality, the SET commands can have multiple that
can even have values other than key-value pairs in it. The most common
of this is SET NAMES and the MariaDB R2DBC connector uses it along with
autocommit=1 to configure itself.

The added test case in sanity_check reproduces the problem where the
query ends up hitting a debug assertion. With this commit applied, the
test passes with no debug assertions.