Use std::make_shared consistently in filters
The use of std::make_shared is better than allocating the filter and
then storing it in the std::shared_ptr as the std::make_shared is able
to allocate both the stored type and the control block for it in one
allocation.

The use of the private constructors in the modules did not bring any
real benefits that would prevent accidental allocation of them. Since
the modules are completely isolated from each other, unless the headers
are explicitly included by something, their allocation can only happen
in the filter that creates them.