Return std::unique_ptr<mxs::Filter> from create()
The return value from the `create()` function is now a smart
pointer. With this and the use of `std::make_shared`, we get rid of most
of the raw calls to `new` in filters. This should make it easier to
search for places where exceptions may cause memory leaks.

The use of `std::make_unique` required that the constructors for the
filter instances are made public. This could also be done by the filter
API directly by constructing the type but unlike in `newSession()`,
there's no instance to interact with when the filter is being created.
Having some way of intercepting the filter instance creation is still a
nice thing to have.