MXS-4783: Clean up opened endpoints if connect() fails
If the router session manages to open the connection to some Endpoints
but then proceeds to throw an exception or return null, the
ServiceEndpoint must clean up after the router and make sure that all
the Endpoints are closed. After the router session has been successfully
created, it will be responsible for closing the connections.