Improve and clean up readwritesplit error messages
The error message about a delayed retry failure caused an extra period
to be added at the end of the error message.

The get_verbose_status() function did not display whether a backend had
failed in a fatal manner or not which makes it hard to figure out the
root cause of a failure.