MXS-5149 Collect line in addition to lineno when ini-parsing
The request for this was a red-herring. The apparent need for
line output was actually caused by MXS-5148, which had nothing
to do with the ini-parsing but with a broken libsnoopy library
on Rocky Linux 8. But collecting the line is nonetheless useful
so now it's done and it is also used in a few places where it
makes sense.