MXS-5258: Prevent retrying of already sent writes
By checking that no responses are expected in can_retry_query(), the
retrying of writes that may cause a double commit is prevented while
still allowing the delaying of the routing to take place that is desired
when primary nodes are not available.

This change implements behavior similar to that of
transaction_replay_safe_commit from 24.08 where delayed_retry is
considered disabled for autocommit statements. However, compared to
transaction_replay_safe_commit, this still allows unsent writes to be
delayed until a write-capable node appears or the time runs out.

Documented the change and explained the reasoning for this change.

Modified the mxs1506_delayed_retry test case to expect the new behavior
of writes not being replayed if they would cause a duplicate write.

Adjusted mxs2609_history_replay to use transactions instead of relying
on the old behavior of writes being retried even if they were
interrupted.