MXS-5259: Retry failed reads on the master
If a read was sent to the master and it fails, it would normally not be
retried unless `delayed_retry` was enabled. This was wrong as the
retrying of reads is controlled by `retry_failed_reads` and the fact
that the query ends up on the master does not always mean that it's a
write.

This fixes a failing test that the fix to MXS-5258 caused.