MXS-5259: Modify test of MXS-2520 to cover MXS-5259
The mxs2520_master_read_reconnect test did not actually test the thing
that it intended to test and simply relied on the side-effect of
delayed_retry being turned on by transaction_replay. The only thing that
is needed is master_reconnection=true and
master_failure_mode=fail_on_write in order to test the retrying of
master reads.

This test case now fails as it reproduces the problem that MXS-5259 is
about.