Print stacktraces from test coredumps
If the coredump happens after the main function has returned, for
example inside an atexit() function, the log will not contain a
stacktrace in which case one has to reproduce the exact environment to
analyze the coredump. Dumping the stacktrace should help avoid that in
the cases where the stacktrace alone is enough to explain where the
crash happens.