Disable JIT if PCRE2 is 10.32 or older
The version of the library on Rocky Linux 8 seems to suffer from
problems in the JIT compilation where the JIT-compiled code reads one
value past the end of the string to be matched. Since the JIT code is
next to impossible to debug without a reproducible test case outside of
MaxScale, the simple approach is to just disable it for old versions as
it's a performance improvement in the first place.