Add WITH_VALGRIND option
Even with the advancements in MSAN usage outside of customized systems,
Valgrind still provides a convenient way to analyze memory related
problems without having to compile MaxScale again.

The WITH_VALGRIND helper option makes the maxscale package depend on
valgrind which removes the need to customize the test system to install
it. It also installs a customized systemd service file that wraps the
MaxScale process in a Valgrind call and configures it correctly for
detecting memory access related errors. This way, a WITH_VALGRIND=Y
build can be run without any changes to the test system.

The valgrind command is set to exit on the first failure which should
cause tests to fail whenever memory access errors are encountered.

Removed the old valgrind mode from the system tests and replaced it with
the new one. The new system should also report valgrind errors if any
were found.