MXS-3760 Wrapping `EntityNode` with a separated component
The `SvgGraphNodes` component, which renders a list of entity nodes,
should be refactored into a separated component to encapsulate related
functionality and event handling. This will reduce the complexity and
responsibilities of the parent component, `EntityDiagram`, making it
more maintainable and easier to test.

This also adds tests for the new component `EntityNodes` and removes
unused helpers in `erdHelper` module.