MXS-5256: Extend SET parser to support more values
The SET statement parser now supports parsing of strings, identifiers
and expressions that use parentheses. With this, the SET statement used
by C/R2DBC is fully parsed and the autocommit assignment is detected
correctly. Extended the unit test to cover the problematic cases.

This commit also fixes the tokenization to take into account assignments
into variables that share a prefix with an expected keywork. For
example, the SESSION_TRACK_SCHEMA was mis-classified as the start of a
SESSION token.

Improved the debug assertion in the client protocol to print out the
type masks if they don't match.