Fix missing supplementary data in `EntityEditorCtr`
A recent cleanup in the `EntityEditorCtr` component on the `develop`
branch caused an issue where the `connData` object was passed to
`schemaInfoService.querySuppData` function. However, the function
expects the connection id via the `connId` attribute instead of `id`.

This addresses that issue and renames the `connData` props in
`TblStructureEditor` to `connReqData` (representing "connection request data")
since `connData` actually refers to the connection object of the
`QueryConn` model.