Revert "MXS-4690: Truncate partially written transactions"
This reverts commit 07ae782f8724e838405e4fdeaa7ef21e6c2cdc2f.

This commit is the cause of MXS-5187. If during a transaction
the master is lost, neither file writing or reading will sync
correctly. This could be handled by brute force connecting file
writing and reading, where the writer dictates when readers can
connect and when they must disconnect, but by writing transactions
to a temp file and then to binlog with appropriate crash recovery
the writer/reader independence is preserved.