MXS-5269: Fail if write to sescmd replier fails
This fixes the problem where the increment of m_expected_responses was
not done if a write to m_current_master failed. The result of this was
that the query was retried and an extra write to m_sescmd_replier was
done. This would cause the debug assertion in client_reply to be hit
that asserts that the value of m_expected_responses is non-negative.