MXS-4783: Deliver lcalls as soon as possible
This makes their behavior a lot more predictable in relation to other
epoll events. With the new approach, the events are delivered as soon as
the epoll handler returns. This guarantees that no other epoll event
will interleave between a failed routeQuery and the corresponding
handleError call. With this, the only thing that routers and filters
have to guarantee is that they do not leak resources and are in a state
that can be destroyed if an exception is thrown in any of the API
functions.

The added cost of this approach is the extra emptiness check for the
lcall vector. The empty() function of a vector is implemented as a
comparison of two pointers being equal which should be relatively
cheap.