Write to m_sescmd_replier first
This way there will be no unnecessary messages written to other nodes if
the write to the backend from which the response is expected fails. It
also clarifies the code so that operations that need to succeed are done
first and the ones that are allowed fail are done later.

Improved the write_session_command() function to log which server is the
replier and removed the unused command byte argument. The function also
now just returns the return value of the write() call as it no longer
has to look at the backend to know whether to treat it as a failure or
not.