MXS-3760 Move pure functions to `diagramUtils`
These functions in `DiagramCtr` always returns the same result when
given the same arguments. Therefore, it's better to move them to
`diagramUtils` to reduce the complexity of the vue component.
In addition, testing pure functions is much easier than testing
them inside vue components, especially when it comes to mocking
stub data.

This also renames some functions to more meaningful names and
repositions them in locations where it improves readability.