MXS-4783: Use lcall() to discard stale connections
Readwritesplit discards stale connections if the targets they connect to
are no longer usable. This must be done outside of the clientReply
function to avoid accidentally deleting the router and filter session
objects while they're still inside their own clientReply calls. This is
similar to deleting the parent node of a tree while the iteration is at
a child node.