MXS-4783: API functions must not be used recursively
A routeQuery() function may only call Endpoint::routeQuery() on one of
its Endpoints, either directly or via the Backend class. Similarly, the
clientReply() and handleError() functions can only cause another call to
the function with the same name to be called. Calling another function
inside of an API function can cause multiple API functions to be entered
at the same time which causes use-after-free issues if the Endpoint gets
closed or internal inconsistencies when the internal status is updated
only after the API function call finishes.

To prevent these sorts of situations, the ServiceEndpoint now asserts
that only one of the API functions is active at any given
time. Additionally, the reference count for the router and filter
sessions should now always be one when the ServiceEndpoint::close()
function is called. This makes the expected behavior more consistent in
that after a call to ServiceEndpoint::close(), all other Targets
underneath it will also be closed immediately.

The behavior of clientReply is also now modified to cause a
handleError() call to be generated if an exception is thrown or the
function returns false. Now all API functions generate a handleError()
call if the API functions fail.

The behavior of the APIs is now documented in the routing.hh and
router.hh headers. The description of the RouterSession class also lists
the common situations that cause problems with the API and how to avoid
them.