MXS-4783: Remove return value of writeq_append()
The appending was always successful. This also now means that
routeQuery() or clientReply() will never return false from
ServiceEndpoint. For ServerEndpoint, the function may still return false
if the connection attempt fails when idle connection pooling is
enabled. This is the first step in potentially removing the return value
from both functions but that is to be done as a separate task outside of
the exception handling changes.