MXS-3760 Remove `disabled` props of `DiagramCtr` component
Instead of disabling certain options in the diagram, it would be
better to show a validation error.

This passes the validate function down to `DiagramCtr` component.
Now, whenever the entity editor is triggered to open, the validate
function is called to ensure that the currently opened editor has
all required fields filled out.