MXS-5289: Update avrorouter parameter documentation
Removed the documentation for the `sorce` parameter as it has been
removed.

The regex parameter documentation uses `none` as the default is no
regular expression, similar to how the `session_trace_match` regular
expression parameters is documented. The ones in kafkacdc don't delcare
a default as they don't have one and an empty string is a valid but
defined regular expression. However, in practice the regular expressions
should be eventually documented to have an empty value as that's how
MaxScale expresses "empty" regular expressions.