MXS-5296 Fix missing default collation value
After refreshing the browser (F5) while the ERD worksheet is opened,
creating a new table for the first time results in a blank collation
field.

This issue occurs because the action to fetch supplementary data,
such as charset and collations, was called only once before the ERD
was generated. All the data is stored in memory, so after refreshing
the browser, this data is lost. To fix this, the fetch action should
also be triggered before creating a new table to ensure the necessary
data is available. There is an empty check in place before the actual
requests are sent, so no redundant requests will be made.