Make mxs2057_systemd_watchdog even slower
Disabling JIT for older PCRE2 versions seems to have sped it up enough
that the test no longer fails. To improve the chances of it failing,
copy the filter a bunch more.

For newer versions of MaxScale where the match limit is lower, a
different approach is needed. The match limit stops the matching in
namedserverfilter and thus the query must be below the limit while still
being long enough to trigger catastrophic backtracing. A query that's
750 characters long seems to be at the sweet spot. When combined with 25
repetitions of the match and with 100 filters, the total execution time
for one query approaches 300 seconds. This should be enough to trigger
the watchdog in all versions.