MXS-5295 Classify SET statements correctly
When classifying a SET statement, the classification was aborted
immediately if NAMES or CHARACTER had been seen, which caused e.g.
"SET autocommit=1,names UTF8MB4" and "SET names UTF8MB4,autocommit=1"
to be classified differently; the former correctly and the latter
incorrectly.

The probable reason for this is that qc_mysqlembedded was broken
and since its result were used for testing qc_sqlite, the latter
became broken as well.

Qc_myselembedded cannot trivially be fixed and since it has
completely been dropped in later MaxScale series, the problematic
test-cases are simply excluded and a specific test case added
for qc_sqlite.