MXS-5302: Exclude binary PS from max_sescmd_history
If a binary protocol prepared statment is not found in MaxScale, it will
end up hitting debug assertions. Unlike text protocol prepared
statements, which are used a lot less, the binary protocol ones are easy
to detect and easier to manage when it comes to excluding them from the
pruning done because of max_sescmd_history.