MXS-5302: Don't prune COM_STMT_PREPARE in mxs::History
The fix needed to be reimplemented in 23.08 as the pruning of the
session command history is now done by the mxs::History class.

The simplest and most straightforward method to do this was to make the
mxs::History class virtual and derive it in the client protocol.The
compiler won't be able to inline the call to can_prune() even if the
MariaDBHistory class is marked as `final`. LTO might be able to inline
it but this is not yet known. All in all, given that history pruning is
a relatively rare event, the overhead of the virtualization should still
be acceptable even if LTO fails to inline it.