MXS-5339: Remove redundant delayed packets
If one server is slower than the rest and the client application uses a
connection pool and closes each prepared statement right after it's
used, the slow server's delayed packet queue will keep growing as long
as the session is alive.

If a COM_STMT_CLOSE packet is delayed and the COM_STMT_PREPARE that it
refers to is also in the delayed packet queue, both of them can be
removed. This helps speed up the execution of session commands on the
slow server as it can safely skip all prepared statements that have been
closed before they are executed.

Although this solves the problem for the common case of one server being
slow, it still suffers from a problem where a client repeatedly sends
COM_STMT_PREPARE packets that the server rejects. To prevent an OOM, a
hard limit that is based on max_sescmd_history is used to forcibly close
any connections that are delayed too much.