MXS-5341: Stop user account managers earlier
The user account manager thread calls
RoutingWorker::execute_concurrently() which may cause a hang if the user
account manager thread happens to do the call after the RoutingWorkers
have stopped but before the stop() function is called in ~Service().

As the user account manager thread depends on RoutingWorkers being up
and running, they must be stopped first during the shutdown.

Removed the obsolete declaraction of service_shutdown() that was not
defined.